requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
plotly>=5.0.0
numpy>=1.21.0 